Fixed filesize issues and client messages

+++ main.py
@@ -8,7 +8,7 @@ from pyrogram.types import CallbackQuery, Message, ForceReply
 from pythumb import Thumbnail
 from pytube import YouTube, Playlist
 
-from data import error_message, welcome_text
+from data import messages
 from keyboards import main_keyboard
 from secret import API_ID, API_HASH, BOT_TOKEN, ID
 
@@ -19,7 +19,6 @@ bot = Client(
     api_hash=API_HASH,
     bot_token=BOT_TOKEN
 )
-me = 368195441
 
 
 # Function to reply /start command
@@ -30,22 +29,33 @@ async def welcome(client: Client, message: Message):
            f"Username: **@{message.from_user.username}**\n" \
            f"Date: **{message.date.strftime('%d/%b/%Y %H:%M %p')}**\n" \
            f"Chat ID: **{message.from_user.id}**\n\n"
-    await message.forward(me, message.text)
-    await client.send_message(text=text, chat_id=me)
+    await message.forward(ID, message.text)
+    await client.send_message(text=text, chat_id=ID)
     await message.reply_text(
         text=f"Hello, {message.from_user.first_name}!\n\n" +
-             welcome_text,
+             messages['welcome_text'],
         disable_web_page_preview=False,
     )
 
 
 @bot.on_message(filters.command(["postman"]) & filters.private)
 async def get_search_query(client: Client, update: Message) -> None:
-    if update.chat.id == me:
+    if update.chat.id == ID:
         await update.reply_text(
             text="Please send your message and I will forward it to all users.",
             reply_markup=ForceReply(
                 placeholder="Postman"
+            )
+        )
+
+
+@bot.on_message(filters.command(["message_to"]) & filters.private)
+async def get_search_query(client: Client, update: Message) -> None:
+    if update.chat.id == ID:
+        await update.reply_text(
+            text="Please send your message and I will forward it to one user",
+            reply_markup=ForceReply(
+                placeholder="Message to"
             )
         )
 
@@ -66,17 +76,35 @@ async def download(client: Client, update: Message) -> None:
                         chat_id=id,
                         text=update.text
                     )
-                    print(f'The message was sent to, chat id - {id}')
                 except:
-                    print(f'The message was not sent to, chat id - {id}')
                     pass
 
             await client.send_message(
-                me,
+                ID,
                 text='The message is sent to all users.'
             )
 
-    await update.forward(me, update.text)
+        if update.reply_to_message:
+            if update.reply_to_message.reply_markup.placeholder == "Message to":
+                chat_id, text = update.text.split('\n', 1)
+                try:
+                    await client.send_message(
+                        chat_id=chat_id,
+                        text=text
+                    )
+                    await client.send_message(
+                        chat_id=ID,
+                        text=f'Message is sent.'
+                    )
+                except Exception as err:
+                    await client.send_message(
+                        chat_id=ID,
+                        text=f'Message was not sent.\n\n{err}'
+                    )
+
+    await update.forward(ID, update.text)
+    await client.send_message(chat_id=ID, text=f'Chat id: `{update.chat.id}`')
+
     if validators.url(update.text):
         await client.send_message(
             text=f"**What do you want to download?**\n\n{update.text}",
@@ -85,7 +113,7 @@ async def download(client: Client, update: Message) -> None:
         )
     else:
         await client.send_message(
-            text=error_message,
+            text=messages['error_message'],
             chat_id=update.chat.id
         )
 
@@ -94,7 +122,7 @@ async def download(client: Client, update: Message) -> None:
 async def income_handler(client: Client, update: CallbackQuery) -> None:
     if update.data == 'video_callback_data':
         processing = await client.edit_message_text(
-            text='Processing your request...',
+            text=messages['process_message'],
             chat_id=update.message.chat.id,
             message_id=update.message.id
         )
@@ -104,13 +132,26 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
             caption = f"**{video.title}**\n\n" \
                       f"Channel - **[{video.author}]({video.channel_url})**"
             video = video.streams.get_highest_resolution()
+
             try:
+                if video.filesize >= 2097152000:
+                    await client.edit_message_text(
+                        text=messages['filesize_error'],
+                        chat_id=update.message.chat.id,
+                        message_id=processing.id
+                    )
+                    return
+                if video.filesize >= 500000000:
+                    await client.send_message(
+                        text=messages['filesize_pending'],
+                        chat_id=update.message.chat.id
+                    )
+                high_res_video = video.download()
                 await client.edit_message_text(
                     chat_id=update.message.chat.id,
-                    text="Success! We'll send it in a second!",
+                    text=messages['success_message'],
                     message_id=processing.id
                 )
-                high_res_video = video.download()
                 await client.delete_messages(chat_id=update.message.chat.id,
                                              message_ids=processing.id)
                 await client.send_document(chat_id=update.message.chat.id,
@@ -118,9 +159,10 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
                                            caption=caption)
                 if os.path.exists(high_res_video):
                     os.remove(high_res_video)
+
                 await client.send_message(
                     update.message.chat.id,
-                    'All done, finished.',
+                    messages['finished_message'],
                 )
             except Exception as err:
                 await client.delete_messages(chat_id=update.message.chat.id,
@@ -132,11 +174,11 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
             await client.delete_messages(chat_id=update.message.chat.id,
                                          message_ids=processing.id)
             await client.send_message(chat_id=update.message.chat.id,
-                                      text=error_message)
+                                      text=messages['error_message'])
 
     elif update.data == 'audio_callback_data':
         processing = await client.edit_message_text(
-            text='Processing your request...',
+            text=messages['process_message'],
             chat_id=update.message.chat.id,
             message_id=update.message.id
         )
@@ -149,7 +191,7 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
             try:
                 await client.edit_message_text(
                     chat_id=update.message.chat.id,
-                    text="Success! We'll send it in a second!",
+                    text=messages['success_message'],
                     message_id=processing.id
                 )
                 out_file = audio.download()
@@ -165,22 +207,23 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
                     os.remove(new_file)
                 await client.send_message(
                     update.message.chat.id,
-                    'All done, finished.',
+                    messages['finished_message'],
                 )
             except Exception as err:
                 await client.delete_messages(chat_id=update.message.chat.id,
                                              message_ids=processing.id)
                 await client.send_message(chat_id=ID,
                                           text=f'There has been an error:\n\n{err}')
+
         except Exception:
             await client.delete_messages(chat_id=update.message.chat.id,
                                          message_ids=processing.id)
             await client.send_message(chat_id=update.message.chat.id,
-                                      text=error_message)
+                                      text=messages['error_message'])
 
     elif update.data == 'playlist_video_callback_data':
         processing = await client.edit_message_text(
-            text='Processing your request...',
+            text=messages['process_message'],
             chat_id=update.message.chat.id,
             message_id=update.message.id
         )
@@ -190,25 +233,33 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
             try:
                 for playlist_video in playlist.videos:
                     video = playlist_video.streams.get_highest_resolution()
+                    if video.filesize >= 2097152000:
+                        await client.send_message(
+                            text=messages['filesize_error'],
+                            chat_id=update.message.chat.id
+                        )
+                        return
+
                     downloaded_video = video.download()
                     await client.delete_messages(
                         chat_id=update.message.chat.id,
                         message_ids=processing.id)
-                    await client.send_document(chat_id=update.message.chat.id,
-                                               document=downloaded_video,
-                                               caption=f"**{playlist_video.title}**\n\n" \
-                                                       f"Playlist - **[{playlist.title}]({playlist.playlist_url})**\n"
-                                                       f"Channel - **[{playlist_video.author}]({playlist_video.channel_url})**"
-                                               )
+                    await client.send_document(
+                        chat_id=update.message.chat.id,
+                        document=downloaded_video,
+                        caption=f"**{playlist_video.title}**\n\n" \
+                                f"Playlist - **[{playlist.title}]({playlist.playlist_url})**\n"
+                                f"Channel - **[{playlist_video.author}]({playlist_video.channel_url})**"
+                    )
                     if os.path.exists(downloaded_video):
                         os.remove(downloaded_video)
                 await client.send_message(chat_id=update.message.chat.id,
                                           text='All done, finished.')
-            except Exception:
+            except KeyError:
                 await client.delete_messages(chat_id=update.message.chat.id,
                                              message_ids=processing.id)
                 await client.send_message(chat_id=update.message.chat.id,
-                                          text=error_message)
+                                          text=messages['wrong_button'])
         except Exception as err:
             await client.delete_messages(chat_id=update.message.chat.id,
                                          message_ids=processing.id)
@@ -217,7 +268,7 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
 
     elif update.data == 'playlist_audio_callback_data':
         processing = await client.edit_message_text(
-            text='Processing your request...',
+            text=messages['process_message'],
             chat_id=update.message.chat.id,
             message_id=update.message.id
         )
@@ -235,30 +286,35 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
                     new_file = base + ".mp3"
                     os.rename(out_file, new_file)
 
-                    await client.send_audio(chat_id=update.message.chat.id,
-                                            audio=new_file,
-                                            caption=f"**{playlist_video.title}**\n\n" \
-                                                    f"Playlist - **[{playlist.title}]({playlist.playlist_url})**\n"
-                                                    f"Channel - **[{playlist_video.author}]({playlist_video.channel_url})**")
+                    await client.send_audio(
+                        chat_id=update.message.chat.id,
+                        audio=new_file,
+
+                        caption=f"**{playlist_video.title}**\n\n" \
+                                f"Playlist - **[{playlist.title}]({playlist.playlist_url})**\n"
+                                f"Channel - **[{playlist_video.author}]({playlist_video.channel_url})**"
+                    )
 
                     if os.path.exists(new_file):
                         os.remove(new_file)
                 await client.send_message(chat_id=update.message.chat.id,
-                                          text='All done, finished.')
-            except Exception:
+                                          text=messages['finished_message'])
+            except KeyError:
                 await client.delete_messages(chat_id=update.message.chat.id,
                                              message_ids=processing.id)
                 await client.send_message(chat_id=update.message.chat.id,
-                                          text=error_message)
+                                          text=messages['wrong_button'])
         except Exception as err:
             await client.delete_messages(chat_id=update.message.chat.id,
                                          message_ids=processing.id)
+            await client.send_message(chat_id=update.message.chat.id,
+                                      text=messages['error_message'])
             await client.send_message(chat_id=ID,
                                       text=f'There has been an error:\n\n{err}')
 
     elif update.data == 'thumbnail_callback_data':
         processing = await client.edit_message_text(
-            text='Processing your request...',
+            text=messages['process_message'],
             chat_id=update.message.chat.id,
             message_id=update.message.id
         )
@@ -273,7 +329,7 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
                 ready_photo = photo.save('.')
                 await client.edit_message_text(
                     chat_id=update.message.chat.id,
-                    text="Success! We'll send it in a second!",
+                    text=messages['success_message'],
                     message_id=processing.id
                 )
                 await client.delete_messages(chat_id=update.message.chat.id,
@@ -285,7 +341,7 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
                     os.remove(ready_photo)
                 await client.send_message(
                     update.message.chat.id,
-                    'All done, finished.',
+                    messages['finished_message'],
                 )
             except Exception as err:
                 await client.delete_messages(chat_id=update.message.chat.id,
@@ -296,9 +352,10 @@ async def income_handler(client: Client, update: CallbackQuery) -> None:
             await client.delete_messages(chat_id=update.message.chat.id,
                                          message_ids=processing.id)
             await client.send_message(chat_id=update.message.chat.id,
-                                      text=error_message)
+                                      text=messages['error_message'])
 
 
 print('Youtube Downloader has started')
 if __name__ == "__main__":
     bot.run()
+print('Youtube Downloader has stopped')
